Moved Rule and Rules types to rules.ts
