+++ src/marked.ts
@@ -140,4 +140,5 @@ export { _Slugger as Slugger } from './Slugger.ts';
 export { _Hooks as Hooks } from './Hooks.ts';
 export { Marked } from './Instance.ts';
 export type * from './MarkedOptions.ts';
+export type * from './rules.ts';
 export type * from './Tokens.ts';
